Some security improvements

+++ usr/share/iserv/www/mod_iserv-nachschreibarbeiten/edit.php
@@ -11,13 +11,17 @@ css_include('style.css');
 PageBlue('Bearbeiten', 'manage');
 
 if(!empty($_POST['action'])) {
-    switch($_POST['action']) { // TODO: Don't we need a userCanEdit() here?
+    switch($_POST['action']) {
         case 'update':
             require_once 'format.inc';
             switch ($_POST['type']) {
                 case 'date':
                     $query = 'UPDATE mod_nachschreibarbeiten_dates SET date=' . qdb($_POST['date']) . ', time=' . qdb($_POST['time']) . ', room=' . qdb($_POST['room']) . ', teacher_act=' . qdb($_POST['teacher']) . ' WHERE id=' . qdb($_POST['id']);
                     $old_date = db_getRow('SELECT * FROM mod_nachschreibarbeiten_dates WHERE id = ' . qdb($_POST['id']));
+                    if(!userCanEdit($old_date)) {
+                        Error('Sie haben keine Berechtigung, diesen Nachschreibtermin zu bearbeiten.');
+                        die;
+                    }
                     if (db_query($query)) {
                         log_insert('Nachschreibtermin geändert. Neue Daten: ' . strip_tags($_POST['date']) . ' um ' . strip_tags($_POST['time']) . ' mit Betreuer_in ' . ActToName(($_POST['teacher'])) . ' in Raum "' . strip_tags($_POST['room']) . '". Alte Daten: ' . getLocalizedFormattedDate($old_date['date'], '%Y-%m-%d') . ' um ' . strip_tags($old_date['time']) . ' mit Betreuer_in ' . ActToName(($old_date['teacher_act'])) . ' in Raum "' . strip_tags($old_date['room']) . '".', null, 'Nachschreibarbeiten');
                         Info('Nachschreibtermin erfolgreich eingetragen.');
@@ -27,6 +31,11 @@ if(!empty($_POST['action'])) {
                     }
                     break;
                 case 'entry':
+                    $old_entry = db_getRow('SELECT * FROM mod_nachschreibarbeiten_entries WHERE id = ' . qdb($_POST['id']));
+                    if(!userCanEdit($old_entry)) {
+                        Error('Sie haben keine Berechtigung, diese Nachschreiber_in zu bearbeiten.');
+                        die;
+                    }
                     $query = 'UPDATE mod_nachschreibarbeiten_entries SET date_id=' . qdb($_POST['date']) . ', student_act=' . qdb($_POST['student']) . ', class=' . qdb($_POST['student_class']) . ', subject=' . qdb($_POST['subject']) . ', additional_material=' . qdb($_POST['additional_material']) . ', duration=' . qdb($_POST['duration']) . ', teacher_act=' . qdb($_POST['teacher']) . ' WHERE id = ' . qdb($_POST['id']);
                     if (db_query($query)) {
                         log_insert('Nachschreiber_in ' . ActToName($_POST['student']) . ' für Termin ' . strip_tags($_POST['date']) . ' geändert.', null, 'Nachschreibarbeiten');
@@ -44,6 +53,11 @@ if(!empty($_POST['action'])) {
         case 'delete':
             switch($_POST['type']) {
                 case 'date':
+                    $old_date = db_getRow('SELECT * FROM mod_nachschreibarbeiten_dates WHERE id = ' . qdb($_POST['id']));
+                    if(!userCanEdit($old_date)) {
+                        Error('Sie haben keine Berechtigung, diesen Nachschreibtermin zu löschen.');
+                        die;
+                    }
                     $query = 'DELETE FROM mod_nachschreibarbeiten_dates WHERE id=' . qdb($_POST['id']);
                     if (db_query($query)) {
                         log_insert('Nachschreibtermin ' . strip_tags($_POST['id']) . ' gelöscht.', null, 'Nachschreibarbeiten');
@@ -54,6 +68,11 @@ if(!empty($_POST['action'])) {
                     }
                     break;
                 case 'entry':
+                    $old_entry = db_getRow('SELECT * FROM mod_nachschreibarbeiten_entries WHERE id = ' . qdb($_POST['id']));
+                    if(!userCanEdit($old_entry)) {
+                        Error('Sie haben keine Berechtigung, diese Nachschreiber_in zu bearbeiten.');
+                        die;
+                    }
                     $query = 'DELETE FROM mod_nachschreibarbeiten_entries WHERE id=' . qdb($_POST['id']);
                     if (db_query($query)) {
                         log_insert('Nachschreiber_in ' . strip_tags($_POST['id']) . ' gelöscht.', null, 'Nachschreibarbeiten');
